changed README.md, refactored code
